chore: 가독성 향상을 위한 수정
- vo의 of를 이용하여 객체를 생성하지 않고 validate 메소드만 사용하여 가독성 및 성능 개선
- int 대신 Math.toIntExact를 이용하여 안전하게 형 변환

+++ src/main/java/com/bannote/userservice/grpcservice/DepartmentGrpcService.java
@@ -117,7 +117,7 @@ public class DepartmentGrpcService extends DepartmentServiceGrpc.DepartmentServi
                     .addAllDepartments(departmentPage.getContent().stream()
                             .map(Department::toProto)
                             .collect(Collectors.toList()))
-                    .setTotalCount((int) departmentPage.getTotalElements())
+                    .setTotalCount(Math.toIntExact(departmentPage.getTotalElements()))
                     .setPage(request.getPage())
                     .setSize(request.getSize())
                     .build();
